trello/requirements.txt
Django==2.2.4
django-cors-headers==3.0.2
django-dotenv==1.4.2
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
psycopg2==2.7.5